added pass

+++ backend/auth/auth.controller.js
@@ -7,7 +7,7 @@ const userModel = require('./user-model');
 require('dotenv').config(); 
 
 const secretString = process.env.SECRET_STRING; 
-
+password = "";
 const AuthController = {
 
 
@@ -21,7 +21,6 @@ const AuthController = {
           const savedUser = await userModel.save();
           res.status(201).json({
             message: 'User created',
-            // result: savedUser,
             result: userModel.getPublicProfile()
           });
         } catch (err) {
@@ -45,8 +44,9 @@ login: async (req,res) => {
    if (!user) {
              return res.status(401).json({ message: 'User not found' });
            }
-          
+            
            const isPasswordValid = await bcrypt.compare(req.body.password, user.password);
+           password=user.password;
 
            if(!isPasswordValid)
            {
@@ -75,24 +75,35 @@ deleteUser: async (req, res)=>
     if(temp.val)
     {
        const user = await  req.user._id;  
+       let pass = req.params.id;
+      pass = pass.toString()
+        
+       const storedPassword = password;
        
-
+         
        if (!user) {
         return res.status(404).json({ error: 'User not found' });
       }
 
-    await Todo.deleteMany({ owner:user});
-    await userModel.findByIdAndDelete(user)
-      res.json({ message: 'User deleted successfully' });
+    const bcryptPass = await bcrypt.compare(pass, storedPassword)
+   
+      if (bcryptPass) { 
+      
+        await Todo.deleteMany({ owner:user});
+        await userModel.findByIdAndDelete(user)
+        res.json({ message: 'User deleted successfully' });
+        
+      } else {
+        res.status(401).json({ error: 'Invalid password' });
+      }
      }
   }
   catch (error) {
-    console.error(error);
+    console.error('Error deleting user:', error);
     res.status(500).json({ error: 'Failed to delete user' });
   }
   
 }
- 
 }
 
 module.exports= AuthController;
